Adds a working version of the evolutionary algorithm
Pretty big commit, adding all the required functions. This version comes with a hardcoded problem instance, just to showcase the algorithm does work SOMETIMES, not ALWAYS though. What is problematic though is the adjustment of parameters. Maybe a change of parameters or the addition of crossover would make the program find the solutions in a better way?
More tests needed to find out